requirements.txt
openai==0.27.8
pydub==0.25.1
python-dotenv==1.0.1
